Move UniqueRef.h to cpp/src/Ice internal header directory.
