+++ cpp/src/Ice/SSL/SSLUtil.cpp
@@ -5,11 +5,11 @@
 #include "SSLUtil.h"
 #include "../Base64.h"
 #include "../Network.h"
+#include "../UniqueRef.h"
 #include "DistinguishedName.h"
 #include "Ice/LocalException.h"
 #include "Ice/SSL/SSLException.h"
 #include "Ice/StringConverter.h"
-#include "Ice/UniqueRef.h"
 #include "IceUtil/Config.h"
 #include "IceUtil/FileUtil.h"
 #include "IceUtil/StringUtil.h"
