+++ cpp/src/Ice/SSL/SecureTransportTransceiverI.h
@@ -9,10 +9,10 @@
 
 #    include "../Network.h"
 #    include "../Transceiver.h"
+#    include "../UniqueRef.h"
 #    include "Ice/Config.h"
 #    include "Ice/SSL/ClientAuthenticationOptions.h"
 #    include "Ice/SSL/ServerAuthenticationOptions.h"
-#    include "Ice/UniqueRef.h"
 #    include "SSLInstanceF.h"
 #    include "SecureTransportEngineF.h"
