+++ cpp/src/Ice/UniqueRef.h
