+++ cpp/src/Ice/ios/StreamConnector.cpp
@@ -8,9 +8,9 @@
 
 #    include "../Network.h"
 #    include "../NetworkProxy.h"
+#    include "../UniqueRef.h"
 #    include "Ice/Exception.h"
 #    include "Ice/Properties.h"
-#    include "Ice/UniqueRef.h"
 #    include "StreamConnector.h"
 #    include "StreamEndpointI.h"
 #    include "StreamTransceiver.h"
